+++ src/main/java/com/royalrangers/repository/RewardRepository.java
@@ -1,4 +1,4 @@
-package com.royalrangers.dao;
+package com.royalrangers.repository;
 
 import com.royalrangers.model.achievement.Reward;
 import org.springframework.data.jpa.repository.JpaRepository;
