+++ src/main/java/com/royalrangers/repository/TaskRepository.java
@@ -1,4 +1,4 @@
-package com.royalrangers.dao;
+package com.royalrangers.repository;
 
 import com.royalrangers.model.achievement.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
