+++ src/main/java/com/royalrangers/repository/TwelveYearAchievementRepository.java
@@ -1,4 +1,4 @@
-package com.royalrangers.dao;
+package com.royalrangers.repository;
 
 import com.royalrangers.model.achievement.TwelveYearAchievement;
 import org.springframework.data.jpa.repository.JpaRepository;
